Metalic background removed!
This is a bug in Visual Studio, and DON'T SAY OTHERWISE AS I WILL DEEM ANYBODY DENYING FACTS A SPAMMER!
I've googled this a lot and one answer about this bug makes even less sense than the other, and since this is code generated by VS itself, and not working, without any changes from my side, you cannot blame me for it not working, no sir!

Then again, Microsoft has always been a bit terrible, when it comes to stuff like this, but this is really beyond logic. And besides, what was the point of having a "visually set up" interface if you have to write tons of additional code just to add a background tile picture into window? (Especially since the previewer just accepts it, which confirms the idiocy behind this even more).